+++ src/app-utils/register-shell-integrations.ts
@@ -15,9 +15,6 @@ export const registerShellIntegrations = (): void => {
 	registerFunctions(
 		{
 			id: 'compose',
-			// TOFIX-SHELL: fix the function type definition
-			// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-			// @ts-ignore
 			fn: openComposerSharedFunction
 		},
 		{